backend/ML/requirements.txt
langdetect
pymorphy2
nltk
pandas
scikit-learn
psycopg2
grpcio 
grpcio-tools 
googleapis-common-protos
python-dotenv
